Try to unify "related" subsampling schemes
This contains a **really hacky** workaround that I hope to resolve with
the Nextstrain team sometime soon.

The idea is that I would create a "dummy" subsampling scheme that
combines both the "early" and the "late" subsampling schemes, and use
that "dummy" as a focus set for the "related" subsampling scheme.
That way, the "related" subsampling schemes would not be biased too much
toward "early" or "late" data.